requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-ckeditor==6.7.0
django-environ==0.10.0
django-js-asset==2.1.0
django-widget-tweaks==1.4.12
Faker==19.3.0
idna==3.4
Pillow==10.0.0
PyAudio==0.2.13
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
sqlparse==0.4.4
termcolor==2.3.0
urllib3==2.0.4
